structural: move library lifecycle events below the service domains
`TMDBSeriesProvider` in `services.metadata` listens to `ScanEndedEvent` and
`RefreshEndedEvent`, so `services.metadata` imported `services.library.events`.
`services.library` in turn injects the metadata provider resolvers and passes
`RemoteSearchResult`, so the two domains formed a package cycle.

An event is a contract between a publisher and listeners that must not know each
other. Housing it in the publisher's domain package reinstates at package level
exactly the coupling the event removes at class level, and it is the listener —
not the publisher — that pays. Move the six library lifecycle events from
`services.library.events` to `services.events.library`, a package below every
service domain that depends on nothing but `java.util`.

Imports only; no event type, field, publisher, listener, or annotation changed.

Library depends on metadata now, and metadata no longer depends back.
