+++ AGENTS.md
@@ -96,7 +96,8 @@ Use Spring's `ApplicationEventPublisher` to decouple side effects from core oper
 - `@TransactionalEventListener(phase = AFTER_COMMIT)` — for publishers inside a transaction; side effects only execute if the transaction commits
 
 **Conventions:**
-- Events are Java records in a `services.<domain>.events` package
+- Events are Java records in a `services.<domain>.events` package while every listener lives in that same domain
+- An event listened to from another service domain moves to `services.events.<domain>` (e.g. `services.events.library`), which sits below all domains — otherwise the listener's domain has to import the publisher's domain, which is the coupling the event was meant to remove, and it forms a package cycle (`ArchitectureTest.serviceSlicesMustBeFreeOfCycles`)
 - Events carry only the data listeners need (IDs, paths) — not full entities
 - Testing: use `CapturingEventPublisher` fake to assert events published; test listeners by direct method invocation
