+++ src/main/java/com/streamarr/server/services/library/LibraryMetadataListener.java
@@ -11,8 +11,8 @@ import com.streamarr.server.domain.LibraryMetadata;
 import com.streamarr.server.jooq.generated.Tables;
 import com.streamarr.server.services.concurrency.MutexFactory;
 import com.streamarr.server.services.concurrency.MutexFactoryProvider;
-import com.streamarr.server.services.library.events.ItemProcessedEvent;
-import com.streamarr.server.services.library.events.ScanCompletedEvent;
+import com.streamarr.server.services.events.library.ItemProcessedEvent;
+import com.streamarr.server.services.events.library.ScanCompletedEvent;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
