+++ src/main/java/com/streamarr/server/services/library/OrphanedMediaFileCleanupService.java
@@ -7,7 +7,7 @@ import com.streamarr.server.repositories.LibraryRepository;
 import com.streamarr.server.repositories.media.MediaFileRepository;
 import com.streamarr.server.services.MovieService;
 import com.streamarr.server.services.filepath.FilepathCodec;
-import com.streamarr.server.services.library.events.ScanCompletedEvent;
+import com.streamarr.server.services.events.library.ScanCompletedEvent;
 import java.nio.file.FileSystem;
 import java.nio.file.Files;
 import java.nio.file.Path;
