+++ src/main/java/com/streamarr/server/services/metadata/series/TMDBSeriesProvider.java
@@ -7,8 +7,8 @@ import com.streamarr.server.domain.Library;
 import com.streamarr.server.domain.media.ContentRating;
 import com.streamarr.server.domain.media.ImageType;
 import com.streamarr.server.domain.media.Series;
-import com.streamarr.server.services.library.events.RefreshEndedEvent;
-import com.streamarr.server.services.library.events.ScanEndedEvent;
+import com.streamarr.server.services.events.library.RefreshEndedEvent;
+import com.streamarr.server.services.events.library.ScanEndedEvent;
 import com.streamarr.server.services.metadata.MetadataResult;
 import com.streamarr.server.services.metadata.RemoteSearchResult;
 import com.streamarr.server.services.metadata.TheMovieDatabaseHttpService;
