+++ src/test/java/com/streamarr/server/services/library/DirectoryWatchingServiceTest.java
@@ -11,8 +11,8 @@ import com.streamarr.server.fakes.FakeLibraryRepository;
 import com.streamarr.server.fixtures.LibraryFixtureCreator;
 import com.streamarr.server.repositories.LibraryRepository;
 import com.streamarr.server.services.filepath.FilepathCodec;
-import com.streamarr.server.services.library.events.LibraryAddedEvent;
-import com.streamarr.server.services.library.events.LibraryRemovedEvent;
+import com.streamarr.server.services.events.library.LibraryAddedEvent;
+import com.streamarr.server.services.events.library.LibraryRemovedEvent;
 import com.streamarr.server.services.validation.IgnoredFileValidator;
 import java.io.IOException;
 import java.nio.file.Files;
