+++ src/test/java/com/streamarr/server/services/library/LibraryManagementServiceTest.java
@@ -54,12 +54,12 @@ import com.streamarr.server.services.PersonService;
 import com.streamarr.server.services.SeriesService;
 import com.streamarr.server.services.concurrency.MutexFactoryProvider;
 import com.streamarr.server.services.filepath.FilepathCodec;
-import com.streamarr.server.services.library.events.ItemProcessedEvent;
-import com.streamarr.server.services.library.events.LibraryAddedEvent;
-import com.streamarr.server.services.library.events.LibraryRemovedEvent;
-import com.streamarr.server.services.library.events.RefreshEndedEvent;
-import com.streamarr.server.services.library.events.ScanCompletedEvent;
-import com.streamarr.server.services.library.events.ScanEndedEvent;
+import com.streamarr.server.services.events.library.ItemProcessedEvent;
+import com.streamarr.server.services.events.library.LibraryAddedEvent;
+import com.streamarr.server.services.events.library.LibraryRemovedEvent;
+import com.streamarr.server.services.events.library.RefreshEndedEvent;
+import com.streamarr.server.services.events.library.ScanCompletedEvent;
+import com.streamarr.server.services.events.library.ScanEndedEvent;
 import com.streamarr.server.services.metadata.MetadataProvider;
 import com.streamarr.server.services.metadata.MetadataResult;
 import com.streamarr.server.services.metadata.RemoteSearchResult;
