+++ src/test/java/com/streamarr/server/services/library/LibraryMetadataListenerIT.java
@@ -10,8 +10,8 @@ import com.streamarr.server.fixtures.LibraryFixtureCreator;
 import com.streamarr.server.repositories.LibraryMetadataRepository;
 import com.streamarr.server.repositories.LibraryRepository;
 import com.streamarr.server.repositories.media.MovieRepository;
-import com.streamarr.server.services.library.events.ItemProcessedEvent;
-import com.streamarr.server.services.library.events.ScanCompletedEvent;
+import com.streamarr.server.services.events.library.ItemProcessedEvent;
+import com.streamarr.server.services.events.library.ScanCompletedEvent;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
