+++ src/test/java/com/streamarr/server/services/library/OrphanedMediaFileCleanupServiceTest.java
@@ -21,7 +21,7 @@ import com.streamarr.server.repositories.media.MovieRepository;
 import com.streamarr.server.services.ImageService;
 import com.streamarr.server.services.MovieService;
 import com.streamarr.server.services.filepath.FilepathCodec;
-import com.streamarr.server.services.library.events.ScanCompletedEvent;
+import com.streamarr.server.services.events.library.ScanCompletedEvent;
 import java.io.IOException;
 import java.nio.file.FileSystem;
 import java.nio.file.Files;
